+++ scripts/results/comparison_results.txt
@@ -1,0 +1,16 @@
+1 Hour Naive
+MSE:1309250987.3530264, MAE: 25646.968138470278, MAPE: 0.021911415695768793
+24 Hour Naive
+MSE:8313610409.492163, MAE: 53561.40178489882, MAPE: 0.04503903611362033
+48 Hour Naive
+MSE:12700080170.195343, MAE: 71828.83552970584, MAPE: 0.060840614639392374
+1 Week Naive
+MSE:3241732518.792032, MAE: 36613.244530062526, MAPE: 0.03159587005190311
+1 Month Naive
+MSE:15661314139.126652, MAE: 88173.07085558205, MAPE: 0.07549169654070388
+1 Year Naive
+MSE:19548581550.618366, MAE: 100630.94241835277, MAPE: 0.08817171023620012
+Neural Network Day 1
+[1037108864.0, 1037108864.0, 22382.91015625, 2.034653663635254]
+Neural Network Day 2
+[2405695744.0, 2405695744.0, 35000.63671875, 3.18095326423645]
